--- requirements.txt
@@ -1,3 +1,0 @@
-greenlet==1.1.2
-SQLAlchemy==1.4.25
-Werkzeug==2.0.1
